requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
youtube-transcript-api
trafilatura
python-dotenv
duckduckgo-search
